Refactors CLI configuration
Overhauls the configuration.
Adds profiles management.
