+++ nextmv/nextmv/cli/main.py
@@ -13,11 +13,14 @@ about the features used here. An example of Rich markup can be found in the
 epilog of the Typer application defined below.
 """
 
+import os
+
 import typer
 from rich import print
+from rich.prompt import Confirm
 
-from nextmv.cli.configure import GO_CLI_PATH, go_cli_exists, load_config, remove_go_cli
-from nextmv.cli.configure import app as configure_app
+from nextmv.cli.configuration.config import CONFIG_DIR, GO_CLI_PATH, load_config
+from nextmv.cli.configuration.configuration import app as configuration_app
 from nextmv.cli.error import error
 from nextmv.cli.version import app as version_app
 
@@ -31,7 +34,7 @@ app = typer.Typer(
 )
 
 # Register subcommands.
-app.add_typer(configure_app)
+app.add_typer(configuration_app, name="configuration")  # Requires a name due to callback in configuration.py.
 app.add_typer(version_app)
 
 
@@ -56,8 +59,9 @@ def handle_go_cli() -> None:
 
     exists = go_cli_exists()
     if exists:
-        delete = typer.confirm(
-            f"Do you want to delete the deprecated Nextmv CLI at {GO_CLI_PATH} now?",
+        delete = Confirm.ask(
+            "Do you want to delete the [italic red]deprecated[/italic red] Nextmv CLI "
+            f"at [italic]{GO_CLI_PATH}[/italic] now?",
             default=True,
         )
         if delete:
@@ -79,10 +83,61 @@ def handle_config_existence(ctx: typer.Context) -> None:
         The Typer context object.
     """
 
-    ignored_commands = {"configure", "version"}
+    ignored_commands = {"configuration", "version"}
     if ctx.invoked_subcommand in ignored_commands:
         return
 
     config = load_config()
     if config == {}:
-        error("No configuration found. Please run [code]nextmv configure[/code].")
+        error("No configuration found. Please run [code]nextmv configuration create[/code].")
+
+
+def go_cli_exists() -> bool:
+    """
+    Check if the Go CLI is installed by looking for the 'nextmv' executable
+    under the config dir.
+
+    Returns
+    -------
+    bool
+        True if the Go CLI is installed, False otherwise.
+    """
+
+    # Check if the Go CLI executable exists
+    exists = GO_CLI_PATH.exists()
+    if exists:
+        print(
+            ":construction: A [italic red]deprecated[/italic red] Nextmv CLI is installed at "
+            f"[italic]{GO_CLI_PATH}[/italic]. You must delete it to avoid conflicts."
+        )
+
+    check_config_in_path()
+
+    return exists
+
+
+def remove_go_cli() -> None:
+    """
+    Remove the Go CLI executable if it exists and notify about PATH cleanup.
+    """
+
+    if GO_CLI_PATH.exists():
+        GO_CLI_PATH.unlink()
+        print(f":white_check_mark: Deleted deprecated {GO_CLI_PATH}.")
+
+    check_config_in_path()
+
+
+def check_config_in_path() -> None:
+    """
+    Check if the configuration directory is in the PATH and notify the user.
+    """
+
+    path_dirs = os.environ.get("PATH", "").split(os.pathsep)
+    config_dir_str = str(CONFIG_DIR)
+
+    if config_dir_str in path_dirs:
+        print(
+            f":construction: [italic]{CONFIG_DIR}[/italic] was found in your [code]PATH[/code]. "
+            f"You should remove any entries related to [italic]{CONFIG_DIR}[/italic] from your [code]PATH[/code]."
+        )
